+++ src/BoundaryConditions/solution_boundary_conditions.jl
@@ -1,3 +1,4 @@
+using Oceananigans.Utils: with_tracers
 using Oceananigans.TurbulenceClosures: with_tracers
 
 """
