+++ src/Diagnostics/cfl.jl
@@ -1,4 +1,4 @@
-using Oceananigans: cell_advection_timescale
+using Oceananigans.Utils: TimeStepWizard
 using Oceananigans.TurbulenceClosures: cell_diffusion_timescale
 
 """
