+++ src/Forcing/model_forcing.jl
@@ -1,3 +1,4 @@
+using Oceananigans.Utils: with_tracers
 using Oceananigans.TurbulenceClosures: with_tracers
 
 """
