+++ src/OutputWriters/jld2_output_writer.jl
@@ -1,7 +1,7 @@
 using Printf
 using JLD2
 
-using Oceananigans: validate_interval
+using Oceananigans.Utils: validate_interval
 
 """
     JLD2OutputWriter{F, I, O, IF, IN, KW} <: AbstractOutputWriter
